Move span classes to separate package
