+++ android/src/main/java/com/expensify/livemarkdown/spans/MarkdownBlockquoteSpan.java
@@ -1,4 +1,4 @@
-package com.expensify.livemarkdown;
+package com.expensify.livemarkdown.spans;
 
 import android.graphics.Canvas;
 import android.graphics.Paint;
