+++ android/src/main/java/com/expensify/livemarkdown/spans/MarkdownFontSizeSpan.java
@@ -1,4 +1,4 @@
-package com.expensify.livemarkdown;
+package com.expensify.livemarkdown.spans;
 
 import android.text.style.AbsoluteSizeSpan;
